requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fastapi==0.135.3
ffmpy==1.0.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gradio==6.11.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
Js2Py==0.74
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
orjson==3.11.8
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pipwin==0.5.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
pyjsparser==2.7.1
pyparsing==3.3.2
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.24
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SpeechRecognition==3.16.0
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
gTTS==2.5.1
